Some Ninja Refactoring: findByCompare
Now all the boilerplate logic is moved into an abstract method that finds
a day by comparing to a previous day. The comparator function is passed
in as an arg. If the comparator returns 1, then the second comparator arg
wins. If the comparator returns -1, then the first comparator arg wins.

This is very similar to how custom sorting works except this has nothing
to do with sorting.