Fix Mocha Test File Pattern
For some reason the file pattern doesn't work as soon as you add a file
at a different level in the file tree. As soon as I added the flattening.test.js
file, the other files that were being found are no longer being found.

Wrapping the file pattern in quotes fixes this for some strange reason.